gemini/agents/market-research-agent/backend/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.15.2
altair==5.4.1
annotated-types==0.7.0
anyio==4.9.0
asyncer==0.0.8
attrs==25.3.0
backoff==2.2.1
bidict==0.23.1
bleach==6.2.0
blinker==1.8.2
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
chevron==0.14.0
click==8.1.8
cloudpickle==3.1.1
colorlog==6.9.0
dataclasses-json==0.6.7
datasets==3.5.0
Deprecated==1.2.14
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.16
dspy==2.6.14
dspy-ai==2.6.17
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
Flask==3.0.3
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.74.0
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-generativeai==0.8.4
google-resumable-media==2.7.2
google_search_results==2.4.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.13.1
grpcio==1.72.0rc1
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
json_repair==0.41.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaggle==1.7.4.2
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-text-splitters==0.3.8
langfuse==2.60.2
langgraph==0.3.30
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.3.31
Lazify==0.4.0
litellm==1.63.7
literalai==0.0.623
magicattr==0.1.6
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.4.3
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.13.3
nest-asyncio==1.6.0
numpy==2.2.4
openai==1.61.0
opentelemetry-api==1.28.0
opentelemetry-exporter-otlp==1.28.0
opentelemetry-exporter-otlp-proto-common==1.28.0
opentelemetry-exporter-otlp-proto-grpc==1.28.0
opentelemetry-exporter-otlp-proto-http==1.28.0
opentelemetry-instrumentation==0.49b0
opentelemetry-proto==1.28.0
opentelemetry-sdk==1.28.0
opentelemetry-semantic-conventions==0.49b0
optuna==4.3.0
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.9.0
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.10.1
python-multipart==0.0.20
python-slugify==8.0.4
python-socketio==5.11.4
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9
setuptools==78.1.0
shapely==2.0.6
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
streamlit==1.40.0
syncer==2.0.3
tenacity==9.1.2
text-unidecode==1.3
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.0.2
tornado==6.4.1
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
uptrace==1.27.0
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.1
watchdog==6.0.0
watchfiles==0.20.0
webencodings==0.5.1
Werkzeug==3.1.3
Wikipedia-API==0.7.1
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.19.0
zipp==3.21.0
zstandard==0.23.0
